Added tracking of import jobs to the welcome page
Added progress tracking for history import jobs to the welcome page. Clicking the "hide" button for a job will now make a PATCH request to the NGA with the URL "/nga/user/<type>/<id>/?show=<true|false>", where <type> is either "export" or "import" and <id> is the job ID (assumed to be unique across imports and exports). Also added option to toggle between showing all jobs or just active jobs (that have not been hidden by the user). The boolean variable "allowHideOnAllJobs" on line 379 can be set to TRUE to show a "Hide" button after all jobs, not just the ones that are finished (or in error/aborted state).